// src/ts/untis/types.ts
export interface ShortData {
    id: number;
    name: string;
    longname: string;
    orgname?: string;
    orgid?: number;
}
// ...
export enum WebUntisElementType {
    CLASS = 1,
    TEACHER = 2,
    SUBJECT = 3,
    ROOM = 4,
    STUDENT = 5,
}

export interface WebElement {
    type: WebUntisElementType;
    id: number;
    orgId: number;
    missing: boolean;
    state: 'REGULAR' | 'ABSENT' | 'SUBSTITUTED';
}

export interface WebElementData extends WebElement {
    element: {
        type: number;
        id: number;
        name: string;
        longName?: string;
        displayname?: string;
        alternatename?: string;
        canViewTimetable: boolean;
        externalKey?: string;
        roomCapacity: number;
    };
}


export interface WebAPITimetable {
    type: "WebAPITimetable";
    id: number;
    lessonId: number;
    lessonNumber: number;
    lessonCode: string;
    lessonText: string;
    periodText: string;
    hasPeriodText: false;
    periodInfo: string;
    periodAttachments: [];
    substText: string;
    date: number;
    startTime: number;
    endTime: number;
    elements: WebElement[];
    studentGroup: string;
    hasInfo: boolean;
    code: number;
    cellState: 'STANDARD' | 'SUBSTITUTION' | 'ROOMSUBSTITUTION' | 'CANCEL';
    priority: number;
    is: {
        roomSubstitution?: boolean;
        substitution?: boolean;
        standard?: boolean;
        event: boolean;
    };
    roomCapacity: number;
    studentCount: number;
    classes: WebElementData[];
    teachers: WebElementData[];
    subjects: WebElementData[];
    rooms: WebElementData[];
    students: WebElementData[];
}